- Refactored DefaultMoveService to add support for getting property data by move Id while staying DRY - Moved controller tests that should have been in api services tests to api services tests - Refactored Knockback parser to eliminate unnecessary branches of logic - Added support to MoveController for getting property data of a specific move by id
